# perf: don't find resource if already matched to previous request URL