Closes #636 

Before:

```
Running 10s test @ http://localhost:1447?test=test
  2 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   141.61ms  175.94ms   1.97s    95.86%
    Req/Sec   424.22     85.40   640.00     73.37%
  8498 requests in 10.10s, 4.51MB read
  Socket errors: connect 0, read 0, write 0, timeout 11
  Non-2xx or 3xx responses: 8498
Requests/sec:    841.39
Transfer/sec:    456.85K
```

After:

```
Running 10s test @ http://localhost:1447
  2 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.46ms    2.91ms 100.13ms   99.54%
    Req/Sec    21.55k     1.24k   22.50k    98.02%
  433043 requests in 10.10s, 56.58MB read
Requests/sec:  42867.70
Transfer/sec:      5.60MB
```