@Guergeiro 2022-05-14
What about [Moogle](https://github.com/drashland/moogle)? Won't it help in anyway?