@crookse 2022-05-14
Moogle would help but the code needed to make this happen is tiny compared to what Moogle has.