@Guergeiro 2022-05-14
And I agree with you. Just checking if it would be better or not.